requirements.txt
google-genai
openai-whisper
gTTS
pydub
sounddevice
numpy
scipy
librosa
opensmile
python-dotenv
pytest